Moved the upload routing/function to its own route file.
